Update Delegation UI Caching strategy
The NextJS caching approach has changed recently.  NextJS now supports
the ability to have `Cache Component`s in addtion to `Server Component`s
and `Client Component`s.  This new feature allows for better and less
"magical" approaches to managing the cached content.  It also allows for
general async functions to be marked and cached where appropriate.
These cached results can be combined to help speed up component
rendering.

These cachable elements are ideal, but by default they populate the data
of components evaluated at `build` time instead of at `runtime`.  Since
we want the service to be able to deliver content with variables
provided at service launch, it requires us to explicitly mark components
/ functions with the `connection` function to ensure that we wait for an
*ACTUAL* connection in order to ensure that this occurs at runtime.

This is due to how NextJS attempts to optimize builds for disementation.
It wants to bundle and build the components at build-time so they are
effectively `static`, which aids in the delivery of the rendered
content.  This makes sense, but it prevents us from being able to modify
or replace information passed to the server at runtime, which is what we
want.  As a result, we are forced to opt-in to dynamic routing, which
prevents us from effectively making this content fully `static`, which
it should effectively be whenever the application is launched.
