--- packages/delegation-ui/src/client_components/__test__/delegation_ui.tsx
@@ -1,9 +1,0 @@
-import { render } from '@testing-library/react';
-import { describe, it } from 'vitest';
-import DelegationUIComponent from '../delegation_ui';
-
-describe('DelegationUI', () => {
-  it('should not throw', () => {
-    render(<DelegationUIComponent />);
-  });
-});
