+++ packages/delegation-ui/src/client_components/delegation_ui.tsx
@@ -1,5 +1,8 @@
 'use client';
 
+import DelegationUIContexts from '@/client_components/provide_contexts';
+import ProvideContextsFromEnv from '@/client_components/provide_contexts_from_env';
+import { EnvironmentConfig } from '@/helpers/read_from_env';
 import {
   IntentCompletedCallbackContext,
   ProvideCappuccinoHotShotQueryServiceAPIContext,
@@ -11,7 +14,34 @@ import {
 } from 'espresso-block-explorer-components/delegation-ui';
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 
-export default function ExplorerClientComponent() {
+export interface DelegationUIClientComponentProps {
+  env: EnvironmentConfig;
+}
+
+/**
+ * DelegationUIClientcomponent is the root client component for the
+ * Delegation UI. It setups the components to provide the relevant
+ * Contexts that are needed for the Delegation UI to work berfore setting
+ * up the other content.
+ */
+export default function DelegationUIClientComponent({
+  env,
+}: DelegationUIClientComponentProps) {
+  return (
+    <ProvideContextsFromEnv env={env}>
+      <DelegationUIContexts>
+        <DelegationUIClientContent />
+      </DelegationUIContexts>
+    </ProvideContextsFromEnv>
+  );
+}
+
+/**
+ * DelegationUIClientContent is responsible for interpretting the query /
+ * search parameters of the URL in order to derive the relevant "intent"
+ * before returning the rest of the Delegation UI.
+ */
+function DelegationUIClientContent() {
   const router = useRouter();
   const pathName = usePathname();
   const params = useSearchParams();
