--- packages/delegation-ui/src/client_components/layout.tsx
@@ -1,113 +1,0 @@
-'use client';
-
-import { EnvironmentProvider } from '@/helpers/environment';
-import { EnvironmentConfig } from '@/helpers/read_from_env';
-import { getWagmiConfigForEnvironment } from '@/helpers/wagmi';
-import { getDefaultConfig, RainbowKitProvider } from '@rainbow-me/rainbowkit';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import {
-  Environment,
-  EnvironmentBanner,
-  ProofOfStakeReleasedContext,
-  ProvideCappuccinoHotShotQueryServiceAPIContext,
-  ProvideCappuccinoNodeValidatorServiceAPIContext,
-  ProvideDerivedDateTimeFormatters,
-  ProvideDerivedNumberFormatters,
-  ProvideESPTokenContract,
-  ProvideL1Methods,
-  ProvideL1ValidatorServiceAPIContext,
-  ProvideLightClientV2Contract,
-  ProvideNavigatorLanguage,
-  ProvideRewardClaimContract,
-  ProvideStakeTableV2Contract,
-  ProvideTickEverySecond,
-  RainbowKitContextInjector,
-} from 'espresso-block-explorer-components';
-import {
-  createConfig,
-  fallback,
-  http,
-  WagmiProvider,
-  type CreateConfigParameters,
-} from 'wagmi';
-
-export interface LayoutClientComponentProps {
-  env: EnvironmentConfig;
-  children: React.ReactNode | React.ReactNode[];
-}
-
-const queryClient = new QueryClient();
-
-export default function LayoutClientComponent({
-  env,
-  children,
-}: LayoutClientComponentProps) {
-  const rawWagmiConfig = getWagmiConfigForEnvironment(
-    env.environment as Environment,
-  );
-
-  const transports = !env.rpc_urls
-    ? {}
-    : {
-        transports: {
-          [rawWagmiConfig.chains[0].id]: fallback(
-            env.rpc_urls.map((url) => http(url)),
-          ),
-        },
-      };
-
-  const rainbowKitWagmiConfig = !env.walletconnect_project_id
-    ? createConfig({
-        ...rawWagmiConfig,
-        ...transports,
-        ssr: true,
-      } as CreateConfigParameters)
-    : getDefaultConfig({
-        appName: 'Espresso Delegation UI',
-        projectId: env.walletconnect_project_id,
-        ...rawWagmiConfig,
-        ...transports,
-        ssr: true,
-      });
-
-  return (
-    <EnvironmentProvider env={env}>
-      <ProofOfStakeReleasedContext.Provider value={env.proof_of_stake_released}>
-        <ProvideNavigatorLanguage>
-          <ProvideDerivedNumberFormatters>
-            <ProvideDerivedDateTimeFormatters>
-              <ProvideTickEverySecond>
-                <ProvideCappuccinoNodeValidatorServiceAPIContext>
-                  <ProvideCappuccinoHotShotQueryServiceAPIContext>
-                    <ProvideL1ValidatorServiceAPIContext>
-                      <WagmiProvider config={rainbowKitWagmiConfig}>
-                        <QueryClientProvider client={queryClient}>
-                          <RainbowKitProvider>
-                            <RainbowKitContextInjector>
-                              <ProvideL1Methods>
-                                <ProvideESPTokenContract>
-                                  <ProvideStakeTableV2Contract>
-                                    <ProvideRewardClaimContract>
-                                      <ProvideLightClientV2Contract>
-                                        <EnvironmentBanner />
-                                        {children}
-                                      </ProvideLightClientV2Contract>
-                                    </ProvideRewardClaimContract>
-                                  </ProvideStakeTableV2Contract>
-                                </ProvideESPTokenContract>
-                              </ProvideL1Methods>
-                            </RainbowKitContextInjector>
-                          </RainbowKitProvider>
-                        </QueryClientProvider>
-                      </WagmiProvider>
-                    </ProvideL1ValidatorServiceAPIContext>
-                  </ProvideCappuccinoHotShotQueryServiceAPIContext>
-                </ProvideCappuccinoNodeValidatorServiceAPIContext>
-              </ProvideTickEverySecond>
-            </ProvideDerivedDateTimeFormatters>
-          </ProvideDerivedNumberFormatters>
-        </ProvideNavigatorLanguage>
-      </ProofOfStakeReleasedContext.Provider>
-    </EnvironmentProvider>
-  );
-}
